# docs/api/openapi.json
{"openapi":"3.0.0","paths":{"/api":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["App"]}},"/api/health":{"get":{"operationId":"HealthController_liveness","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/api/health/ready":{"get":{"operationId":"HealthController_readiness","parameters":[],"responses":{"200":{"description":""}},"tags":["Health"]}},"/api/auth/register":{"post":{"operationId":"AuthController_register","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDto"}}}},"responses":{"201":{"description":"Compte créé"},"409":{"description":"Email déjà utilisé"}},"summary":"Créer un compte","tags":["auth"]}},"/api/auth/login":{"post":{"operationId":"AuthController_login","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginDto"}}}},"responses":{"200":{"description":"Jeton JWT émis"},"401":{"description":"Identifiants invalides"}},"summary":"Se connecter","tags":["auth"]}},"/api/auth/me":{"get":{"operationId":"AuthController_me","parameters":[],"responses":{"200":{"description":"Identité résolue depuis le JWT"},"401":{"description":"JWT manquant ou invalide"}},"security":[{"bearer":[]}],"summary":"Utilisateur courant","tags":["auth"]}},"/api/files/uploads":{"post":{"operationId":"FilesController_initiate","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/InitiateUploadDto"}}}},"responses":{"201":{"description":"fileId, taille de partie et URLs pré-signées"},"400":{"description":"Extension interdite ou tags en double"},"413":{"description":"Taille déclarée supérieure à 1 Gio"}},"security":[{"bearer":[]}],"summary":"Démarrer un téléversement multipart","tags":["files"]}},"/api/files/uploads/{id}/parts":{"get":{"operationId":"FilesController_getParts","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Parties complétées et URLs des parties manquantes"},"400":{"description":"id mal formé (pas un UUID)"},"404":{"description":"Envoi inconnu ou n'appartenant pas à l'utilisateur"},"410":{"description":"Multipart expiré côté stockage"}},"security":[{"bearer":[]}],"summary":"Lister les parties déjà envoyées et re-signer les manquantes","tags":["files"]}},"/api/files/uploads/{id}/status":{"get":{"operationId":"FilesController_getStatus","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"État courant ; le jeton n'est inclus que dans l'état `ready`"},"400":{"description":"id mal formé (pas un UUID)"},"404":{"description":"Envoi inconnu ou n'appartenant pas à l'utilisateur"}},"security":[{"bearer":[]}],"summary":"Suivre l'état d'un fichier pendant l'analyse","tags":["files"]}},"/api/files/uploads/{id}/complete":{"post":{"operationId":"FilesController_complete","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CompleteUploadDto"}}}},"responses":{"200":{"description":"Fichier accepté, analyse en file d'attente (aucun lien avant l'état `ready`)"},"400":{"description":"id mal formé (pas un UUID)"},"404":{"description":"Envoi inconnu ou n'appartenant pas à l'utilisateur"},"409":{"description":"L'envoi n'est plus en attente"},"413":{"description":"Taille réelle supérieure à 1 Gio"}},"security":[{"bearer":[]}],"summary":"Finaliser un téléversement multipart","tags":["files"]}},"/api/files/uploads/{id}":{"delete":{"operationId":"FilesController_abort","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Envoi annulé"},"400":{"description":"id mal formé (pas un UUID)"},"404":{"description":"Envoi inconnu ou n'appartenant pas à l'utilisateur"},"409":{"description":"L'envoi n'est plus en attente"}},"security":[{"bearer":[]}],"summary":"Annuler un téléversement multipart","tags":["files"]}},"/api/files/{id}":{"delete":{"description":"Supprime la ligne et, si un objet existe encore en stockage, l'objet (ou avorte le multipart s'il est encore en vol). Idempotent : un second appel sur un fichier déjà supprimé renvoie 404, jamais 500.","operationId":"FileDeletionController_remove","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"204":{"description":"Fichier supprimé"},"400":{"description":"id mal formé (pas un UUID)"},"404":{"description":"Fichier inconnu ou n'appartenant pas à l'utilisateur"}},"security":[{"bearer":[]}],"summary":"Supprimer un fichier (Mon espace)","tags":["files"]}},"/api/files":{"get":{"operationId":"FileHistoryController_list","parameters":[{"name":"filter","required":false,"in":"query","description":"Tous/Actifs/Expiré. « all » inclut aussi rejected, qui n'a pas son propre onglet.","schema":{"default":"all","type":"string","enum":["all","active","expired"]}}],"responses":{"200":{"description":"Fichiers ready/expired/rejected, filtrés au propriétaire, du plus récent au plus ancien"}},"security":[{"bearer":[]}],"summary":"Historique des fichiers du propriétaire (Mon espace)","tags":["files"]}},"/api/d/{token}":{"get":{"operationId":"DownloadController_getMetadata","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Fichier prêt ; URL incluse si aucun mot de passe"},"202":{"description":"Analyse en cours"},"404":{"description":"Lien invalide (jeton inconnu, refusé ou abandonné — réponse identique dans les trois cas)"},"410":{"description":"Fichier expiré"}},"summary":"Métadonnées avant téléchargement","tags":["download"]},"post":{"operationId":"DownloadController_verifyPassword","parameters":[{"name":"token","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/VerifyPasswordDto"}}}},"responses":{"200":{"description":"URL pré-signée, valable 60 secondes"},"401":{"description":"Mot de passe incorrect"},"404":{"description":"Lien invalide"},"410":{"description":"Fichier expiré ou toujours en analyse"}},"summary":"Vérifier le mot de passe et obtenir une URL de téléchargement","tags":["download"]}}},"info":{"title":"DataShare API","description":"Contrat OpenAPI généré depuis les DTO de validation.","version":"0.0.1","contact":{}},"tags":[],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http"}},"schemas":{"RegisterDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string","minLength":8},"displayName":{"type":"string","maxLength":100}},"required":["email","password"]},"LoginDto":{"type":"object","properties":{"email":{"type":"string","format":"email"},"password":{"type":"string"}},"required":["email","password"]},"InitiateUploadDto":{"type":"object","properties":{"expiresInDays":{"type":"number","minimum":1,"maximum":7,"default":7},"password":{"type":"string","minLength":6},"tags":{"maxItems":20,"type":"array","items":{"type":"string","maxLength":30}},"showSender":{"type":"boolean","default":false},"originalName":{"type":"string","maxLength":255},"mimeType":{"type":"string","maxLength":127},"sizeBytes":{"type":"number","minimum":1}},"required":["originalName","mimeType","sizeBytes"]},"UploadedPartDto":{"type":"object","properties":{"partNumber":{"type":"number","minimum":1},"etag":{"type":"string"}},"required":["partNumber","etag"]},"CompleteUploadDto":{"type":"object","properties":{"parts":{"minItems":1,"type":"array","items":{"$ref":"#/components/schemas/UploadedPartDto"}}},"required":["parts"]},"VerifyPasswordDto":{"type":"object","properties":{"password":{"type":"string"}}}}}}